+++ examples/App.vue
@@ -1,16 +1,24 @@
 <script setup>
-import { StarrySky,DigitalScroll, AlanButton,CodeBackgroundWall,AlanBackTop } from '../packages/index';
+import { StarrySky, DigitalScroll, AlanButton, CodeBackgroundWall, AlanBackTop, TherMometer } from '../packages/index';
 
-const goTopCompleteCb = () => {}
+const goTopCompleteCb = () => { }
 </script>
 
 <template>
-  <div class="h"></div>
-	<starry-sky :stars-count="888" :distance="1000" />
-	<!-- <digital-scroll :targetNumber="1234" targetClass="class"></digital-scroll>
+  <starry-sky :stars-count="888"
+              :distance="1000" />
+  <TherMometer :show="true"
+               :data="{
+                 max: 50,
+                 min: 0,
+                 value: 60,
+                 warn: 10,
+               }" />
+  <!-- <digital-scroll :targetNumber="1234" targetClass="class"></digital-scroll>
   <alan-button disabled  type="primary">111</alan-button>
   <CodeBackgroundWall></CodeBackgroundWall> -->
-  <AlanBackTop  @goTopCompleteCb="goTopCompleteCb"></AlanBackTop>
+  <AlanBackTop @goTopCompleteCb="goTopCompleteCb"></AlanBackTop>
+  <div class="h"></div>
 </template>
 
 <style scoped>