+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## <small>1.0.19 (2022-10-26)</small>
+## <small>1.0.20 (2022-10-26)</small>
 
 * ✨新增: index.html 格式调整 ([f769afa](https://github.com/2401345934/vue3-alan-vite-component/commit/f769afa))
 * ✨新增(custom): 修改 vite 配置 ([0ccb107](https://github.com/2401345934/vue3-alan-vite-component/commit/0ccb107))
@@ -13,6 +13,7 @@
 * ✨新增(packages): 新增 prepublish 每次 publish 之前自动 build ([1fa3129](https://github.com/2401345934/vue3-alan-vite-component/commit/1fa3129))
 * ✨新增(packages): 新增BackTop ([b9a9f70](https://github.com/2401345934/vue3-alan-vite-component/commit/b9a9f70))
 * ✨新增(packages): codeBackgroundWall 背景墙组件 ([81e3fa6](https://github.com/2401345934/vue3-alan-vite-component/commit/81e3fa6))
+* ✨新增(packages): monorepo升级 ([e1ea8e8](https://github.com/2401345934/vue3-alan-vite-component/commit/e1ea8e8))
 * ✨新增(packages/DigitalScroll): 引入 ts ([9579d53](https://github.com/2401345934/vue3-alan-vite-component/commit/9579d53))
 * 🐛修复(custom): 修复 button bug ([f4713d3](https://github.com/2401345934/vue3-alan-vite-component/commit/f4713d3))
 * 🐛修复(package): 修复打包后其他项目引入报错 ([6dcc934](https://github.com/2401345934/vue3-alan-vite-component/commit/6dcc934))